Move SDD package to better location
