--- src/tarski/sdd/__init__.py
@@ -1,2 +1,0 @@
-
-from .sdd import process_problem
